backend-python/requirements.txt
torch
torchvision
torchaudio
setuptools==69.5.1
rwkv==0.8.26
langchain==0.0.322
fastapi==0.109.1
uvicorn==0.23.2
sse-starlette==1.6.5
pydantic==2.4.2
psutil==5.9.6
gputil==1.4.0
tiktoken==0.5.1
ftfy==6.1.1
lm-dataformat==0.0.20
numpy==1.24.4
tokenizers==0.14.1
tqdm==4.66.1
midi2audio==0.1.1
mido==1.3.0
safetensors==0.4.0
PyMuPDF==1.23.5
python-multipart==0.0.7
Cython==3.0.4
cyac==1.9
torch-directml==0.1.13.1.dev230413
